refactor twitter example
